\name{USPS-internal}
\alias{USPS-internal}
\alias{plot.SPSbalan}
\alias{plot.SPSoutco}
\alias{plot.UPShclus}
\alias{plot.UPSivadj}
\alias{plot.UPSaltdd}
\alias{print.SPSbalan}
\alias{print.SPSloess}
\alias{print.SPSlogit}
\alias{print.SPSoutco}
\alias{print.SPSsmoot}
\alias{print.UPShclus}
\alias{print.UPSivadj}
\alias{print.UPSnnltd}
\alias{print.UPSaltdd}
\alias{UPlinint}
\title{Internal USPS functions.}
\description{Internal method functions for the USPS package.  The functions except UPlinint
 require a USPS object as their first argument, but accept optional arguments to plot() or
 print() documented elsewhere.}
\author{Bob Obenchain <wizbob@att.net>}
\keyword{methods}
\keyword{internal}