\name{USPS-internal}
\alias{USPS-internal}
\alias{plot.SPSbalan}
\alias{plot.SPSoutco}
\alias{plot.UPShclus}
\alias{plot.UPSivadj}
\alias{print.SPSbalan}
\alias{print.SPSloess}
\alias{print.SPSlogit}
\alias{print.SPSoutco}
\alias{print.SPSsmoot}
\alias{print.UPShclus}
\alias{print.UPSivadj}
\alias{print.UPSnnltd}
\title{Internal USPS functions.}
\description{Internal method functions for the USPS package.  The functions require a USPS object as their first
 argument, but accept no other arguments ...except optional arguments to plot() or print() that are documented elsewhere.}
\author{Bob Obenchain <sunsetstats@earthlink.net>}
\keyword{methods}
\keyword{internal}